set strategy to 'recreate'
because znc is a single instance service, the normal rolling update
process when updating the deployment won't work: the recreate pod
will fail because the active pod has the znc configuration locked.

by using the 'recreate' strategy, the active pod is immediately
replaced by the new pod. See [1] for some discussion on this topic.

[1]: https://kubernetes.io/docs/tasks/run-application/run-single-instance-stateful-application/